\name{cement}
\alias{cement}
\alias{clothes}
\alias{oil}
\alias{ore}
\alias{plastic}
\title{Industrial Production of China}
\format{Each time series is an object of class \code{"ts"}.}
\source{
National Bureau of Statistics of China
\url{http://www.stats.gov.cn/english/}
}
\description{
Industrial Production of China
}
\keyword{datasets}

