\name{cntrade}
\alias{clothes}
\alias{cntrade}
\alias{oil}
\alias{ore}
\alias{plastic}
\title{Exports and Imports of China}
\format{Each time series is an object of class \code{"ts"}.}
\source{
China Customs \url{http://www.stats.gov.cn/english/}
}
\description{
Monthly exports and imoports of China from July 1983 to
December 2013.
}
\details{
In 100 mio. U.S. Dollar.
}
\keyword{datasets}

