% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal-package.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Consumer Price Index of Switzerland}
\format{
Time series of class \code{"ts"}.
}
\source{
Federal Statistical Office, Switzerland
}
\description{
Monthly consumer price index of Switzerland. Base year is 1993.
}
\examples{

data(seasonal)
cpi
}
\keyword{datasets}
