% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fivebestmdl}
\alias{fivebestmdl}
\title{Five Best ARIMA Models}
\usage{
fivebestmdl(x)
}
\arguments{
\item{x}{object of class \code{"seas"}}
}
\description{
Returns the five best models as chosen by the BIC criterion. It needs the
\code{automdl} spec to be activated (default). If it is not activated, the
function tries to re-evaluate the model with the \code{automdl} spec
activated.
}
\examples{
\donttest{

m <- seas(AirPassengers)
fivebestmdl(m)
}
}
\references{
Vignette with a more detailed description:
\url{https://www.seasonal.website/seasonal.html}

Comprehensive list of R examples from the X-13ARIMA-SEATS manual:
\url{https://www.seasonal.website/examples.html}

Official X-13ARIMA-SEATS manual:
\url{https://www2.census.gov/software/x-13arima-seats/x13as/unix-linux/documentation/docx13as.pdf}
}
\seealso{
\code{\link[=seas]{seas()}} for the main function.

\code{\link[=series]{series()}}, for universal X-13 output extraction.

\code{\link[=plot.seas]{plot.seas()}}, for diagnostical plots.

\code{\link[=out]{out()}}, for accessing the full output of X-13ARIMA-SEATS.
}
