% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.spc.R
\name{import.spc}
\alias{import.spc}
\alias{print.import.spc}
\title{Import X-13 \code{.spc} Files (experimental)}
\usage{
import.spc(file)

\method{print}{import.spc}(x, ...)
}
\arguments{
\item{file}{character, name of the X-13 \code{.spc} file}

\item{x}{object of class \code{import.spc}}

\item{...}{further arguments, not used}
}
\value{
returns an object of class \code{import.spc}, which is a list with the following (optional) objects of class \code{call}:
  \item{call}{the call to \code{\link{seas}}}
  \item{x}{the call to retrieve the data for the input series} 
  \item{xtrans}{the call to retrieve the data for the \code{xtrans} series (if required by the call)} 
  \item{xreg}{the call to retrieve the data for the \code{xreg} series (if required by the call)}
}
\description{
Utility function to import \code{.spc} files from X-13. It generates a list
of calls to \code{seas} (and  \code{import.ts}) that can be run in R.
Evaluating these calls should perform the same X-13 run as the original
\code{.spc} file. The \code{print} method displays the calls in a way that 
they can be copy-pasted into an R script.
}
\examples{

# importing the orginal X-13 example file
import.spc(system.file("tests", "Testairline.spc", package="seasonal"))

\dontrun{

### reading .spc with multiple user regression and transformation series

# running a complex seas call and save output in a temporary directory
tdir <- tempdir()
seas(x = AirPassengers, xreg = cbind(a = genhol(cny, start = 1, end = 4,
    center = "calendar"), b = genhol(cny, start = -3, end = 0,
    center = "calendar")), xtrans = cbind(sqrt(AirPassengers), AirPassengers^3), 
    transform.function = "log", transform.type = "temporary", 
    regression.aictest = "td", regression.usertype = "holiday", dir = tdir, 
    out = TRUE)

# importing the .spc file from the temporary location
ll <- import.spc(file.path(tdir, "iofile.spc"))

# ll is list containing 4 calls: 
# - 3 calls to import.ts(), which read the series from the X-13 data files
# - and 1 call to seas() which performs the seasonal adjustment in R
str(ll)

# to replicate the original X-13 operation, run all 4 calls. You can either
# copy/paste and run the print() output:
ll

# or use eval() to evaluate the call(s). To evaluate the first call and
# import the x variable:
eval(ll$x)

# to run all 4 calls, use lapply() and eval() to evaluate all calls in the
# list:
ee <- lapply(ll, eval, envir = globalenv())
ee$call  # the 'seas' object, produced by the final call to seas()
}
}
\seealso{
\code{\link{import.ts}}, for importing X-13 data files.

\code{\link{seas}} for the main function of seasonal.
}

