% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{predict.seas}
\alias{predict.seas}
\title{Predict Method}
\usage{
\method{predict}{seas}(object, newdata)
}
\arguments{
\item{newdata}{an object of class \code{"ts"}}

\item{an}{object of class \code{"seas"}, usually, a result of a 
call to \code{\link{seas}}.}
}
\description{
Predict Method
}

