% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.spc.R
\name{read.ts}
\alias{read.ts}
\title{read time series from X-13 input files}
\usage{
read.ts(file, format = "datevalue", start = NULL, frequency = NULL)
}
\arguments{
\item{x}{dsfdf}
}
\description{
helper function to read time series from X-13 input files
}
\examples{
read.ts("/Users/christoph/tmp/data.dta")
}

