% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.seas.R
\name{robust.seas}
\alias{robust.seas}
\title{A Robust Version of \code{seas} that \emph{Always} Works (Experimental)}
\usage{
robust.seas(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{seas}}.}
}
\value{
returns an object of class \code{"seas"}.
}
\description{
\code{robust.seas} has exactly the same usage as \code{\link{seas}}, the main 
function of seasonal, but will always work. If \code{seas} is failing, 
\code{robust.seas} tries an an alternative specification, and returns a 
message. `robust.seas` currently works with all 3000+ series of the M3 
forecast competition.
}
\details{
If \code{seas} fails, \code{robust.seas} tries the following:

\enumerate{
\item adjusting invalid start year
\item increasing \code{maxiter} to 10000
\item fix model to (0 1 1)(0 1 1)
\item turn off AIC testing
\item turn off outlier detection
}

If \code{robust.seas} is applied to annual series (in which case you do not
need seasonal adjustment), it will proceed slightly different:

\enumerate{
\item adjusting invalid start year
\item ensure SEATS, X11 and AIC testing are off
\item turn off outlier detection
\item fix model to (1 1 0)
}

If it still fails, it will return an error, along with the suggestion to post the example on:
\url{https://github.com/christophsax/seasonal/wiki/Breaking-Examples-(and-Possible-Solutions)}.
}
\examples{
\dontrun{
 
x <- ts(1:40)  # an annual time series with an invalid start year

# not working
seas(x)

# working
robust.seas(x)
}
}
\seealso{
\code{\link{seas}} for the main function of seasonal.
}

