% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{arimamodel}
\alias{arimamodel}
\title{Defunct Functions}
\usage{
arimamodel(x)
}
\arguments{
\item{x}{object of class \code{"seas"}}
}
\description{
The \code{arimamodel} functions is defunct now. Use 
the more univeral \code{\link{udg}} function.
}
\examples{
\dontrun{
m <- seas(AirPassengers)
udg(x, "x13mdl")
}
}
\seealso{
\code{\link{udg}}, for universal access to X-13 statistics
}

