% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_list.R
\docType{data}
\name{series_list}
\alias{series_list}
\title{List of All Possible Series That Can Be Extracted via the \code{series} Function}
\format{
An object of class \code{list} of length 210.
}
\usage{
series_list
}
\description{
This list is intended as a helper when using autocompletion during calls
to \code{series} (see examples).
}
\examples{
m <- seas(AirPassengers)
series(m, series_list$seats.seasonal)
}
\keyword{datasets}
