% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal-package.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{United States Unemployment Level}
\format{Each time series is an object of class \code{"ts"}.}
\source{
U.S. Bureau of Labor Statistics, retrieved from FRED, Federal Reserve
  Bank of St. Louis; \url{https://fred.stlouisfed.org/series/LNU03000000},
  December 14, 2016.
}
\description{
Thousands of Persons
}
\examples{

data(seasonal)
unemp
}
\keyword{datasets}
