% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.seas}
\alias{update.seas}
\title{Update Method}
\usage{
\method{update}{seas}(object, ...)
}
\arguments{
\item{...}{spec-arguments options sent to X-13ARIMA-SEATS (see seas)}

\item{an}{object of class \code{"seas"}, usually, a result of a 
call to \code{\link{seas}}.}
}
\description{
Update Method
}

