% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.seas}
\alias{as.data.frame.seas}
\alias{as.data.frame.summary.seas}
\alias{as.data.frame.udg}
\title{Coerce Output to \code{data.frame}}
\usage{
\method{as.data.frame}{seas}(x, ...)

\method{as.data.frame}{udg}(x, ...)

\method{as.data.frame}{summary.seas}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"seas"}, \code{"udg"} or \code{"summary.seas"}, usually, a result of a 
call to the functions with the same name.}

\item{...}{unused.}
}
\value{
a \code{data.frame} without row names.
}
\description{
These methods coerce the output to a \code{data.frame}. This is useful for
further processing. (This is a second attempt of doing that; the first
experimenal attempt in version 1.4 used an argument \code{data.frame} in the
call to the functions, and is now obsolete. The present approach seems
cleaner, but still consider it as \strong{experimental}.)
}
\details{
The data.frames produced by these functions follow the nameing conventions
proposed by the broom package, but do not rely on the package otherwise.
}
\examples{
\dontrun{
m <- seas(AirPassengers, x11 = "")

# a data.frame containing the M-statistics
as.data.frame(udg(m, paste0("f3.m", formatC(1:10, width=2, flag="0"))))

# a data.frame containing data
as.data.frame(m)

# a data.frame containing the summary information on the coefficients
as.data.frame(summary(m))
}
}

