% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal-package.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Consumer Price Index of Switzerland}
\format{Time series of class \code{"ts"}.}
\source{
Federal Statistical Office, Switzerland 
\url{http://www.bfs.admin.ch/bfs/portal/de/index/themen/05/02/blank/key/basis_aktuell.html} 
(cc-d-05.02.12)
}
\description{
Monthly consumer price index of Switzerland. Base year is 1993.
}
\examples{

data(seasonal)
cpi
}
\keyword{datasets}
