% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal-package.R
\docType{data}
\name{exp}
\alias{exp}
\alias{imp}
\title{Exports and Imports of China}
\format{
Each time series is an object of class \code{"ts"}.
}
\source{
China Customs
}
\description{
Monthly exports and imports of China (July 1983 to December 2013).
}
\details{
In 100 mio U.S. Dollar.
}
\examples{

data(seasonal)
imp
exp
}
\keyword{datasets}
