% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iip}
\alias{iip}
\title{Industrial Production of India}
\format{Time series of class \code{"ts"}.}
\source{
Central Statistics Office of the Ministry of Statistics and Programme
Implementation
\url{http://mospi.nic.in/Mospi_New/upload/iip/table3_04_05.xls}, with help
from Pinaki Mukherjee
}
\description{
Industrial Production of India (IIP).
}
\details{
Index value. IIP is used for measuring the performance overall industrial
sector of the Indian economy. IIP is compiled by using data from 16 source
agencies.
}
\examples{
data(seasonal)
iip
}
\keyword{datasets}

