% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.spc.R
\name{import.spc}
\alias{import.spc}
\title{read time series from X-13 input files}
\usage{
import.spc(file)
}
\arguments{
\item{x}{dsfdf}
}
\description{
helper function to read time series from X-13 input files
}
\examples{
import.spc("~/Desktop/Testairline.spc")
import.spc("/Users/christoph/tmp/iofile.spc")
}

