% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.spc.R
\name{import.ts}
\alias{import.ts}
\title{Import Series from X-13 Data Files (experimental)}
\usage{
import.ts(file, format = "datevalue", start = NULL, frequency = NULL,
  name = NULL)
}
\arguments{
\item{file}{character, name of the X-13 file which the data are to be read from}

\item{format}{a valid X-13 file format as described in 7.15, p. 173 of the
X-13 manual.}

\item{start}{vector of length 2, time of the first observation (only for
formats \code{"free"} and \code{"freecomma"})}

\item{frequency}{the number of observations per unit of time (only for
formats \code{"free"} and \code{"freecomma"})}

\item{name}{name of the series, to select from a multiple time series. For
compatibility with the existing spc files in the x11 format. Omit if you
want to read all time series from a file.}
}
\description{
Helper function to read time series from X-13 input files. A call to
\code{import.ts} is constructed and included in the output of
\code{\link{import.spc}}.
\bold{This is an experimental function that may change substantially in the future.}
}
\examples{
tpath <- file.path(path.package("seasonal"), "tests")

import.ts(file.path(tpath, "datavalue1.dta"))
import.ts(file.path(tpath, "free1.txt"), format = "free", start = c(1949, 1),
          frequency = 12)
import.ts(file.path(tpath, "free2.txt"), format = "free", start = c(1949, 1),
          frequency = 12)
}

