% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Outlier Time series}
\usage{
outlier(x, full = FALSE)
}
\arguments{
\item{x}{an object of class \code{"seas"}.}

\item{full}{logical, should the full label of the outlier be shown? If
  \code{FALSE}, only the type of the outlier is shown.}
}
\value{
character string time series with outliers.
}
\description{
Returns an object of class \code{"ts"} that contains the names of the
outliers.
}
\examples{
\dontrun{
x <- seas(AirPassengers)
outlier(x)
}
}

