% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{slidingspans}
\alias{plot.revisions}
\alias{plot.slidingspans}
\alias{regressioneffects}
\alias{revisions}
\alias{slidingspans}
\title{Defunct Functions}
\usage{
slidingspans(x, ...)

regressioneffects(x)

revisions(x, ...)

\method{plot}{revisions}(x, ...)

\method{plot}{slidingspans}(x, ...)
}
\arguments{
\item{x}{object of class \code{"seas"}}

\item{...}{further arguments}
}
\description{
Functions are defunct. Use the more univeral series function.
}
\examples{
\dontrun{
# history spec (replaces the defunct 'revisions' function)
series(m, "history.trendestimates")
series(m, "history.sfestimates")
series(m, "history.saestimates")
series(m, c("history.sfestimates", "history.trendestimates"))

# slidingspans spec (replaces the defunct 'slidingspans' function)
series(m, "slidingspans.sfspans")
series(m, "slidingspans.tdspans")

# regressioneffects
series(m, "estimate.regressioneffects")
}
}
\seealso{
\code{\link{series}}, for universal X-13 output extraction.
}

