% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{inspect}
\alias{inspect}
\title{Deprecated Functions}
\usage{
inspect(x, fun = NULL, check.version = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"seas"}.}

\item{fun}{a function or a list of functions (see details)}

\item{check.version}{logical, should the version of shiny be checked}

\item{quiet}{logical, if \code{TRUE} (default), error messages from calls in 
inspect are not shown in the console}

\item{...}{further arguments, passed on to 
\code{\link[shiny]{runApp}}. (The \code{launch.browser} argument of 
version 0.8 can be still used that way)}
}
\description{
The \code{inspect} functions is deprecated and will be defunct soon. Use 
the extended \code{\link{view}} function instead.
}
\examples{
\dontrun{
m <- seas(AirPassengers)
view(m)
}
}
\seealso{
\code{\link{view}}, for an extended graphical user interface.
}

