\docType{package}
\name{seasonal-package}
\alias{seasonal}
\alias{seasonal-package}
\title{seasonal: R interface to X-13ARIMA-SEATS}
\description{
seasonal is an easy-to-use R-interface to X-13ARIMA-SEATS,
a seasonal adjustment software developed by the United
States Census Bureau. X-13ARIMA-SEATS combines and extends
the capabilities of the older X-12ARIMA (developed by the
Census Bureau) and the TRAMO-SEATS (developed by the Bank
of Spain) software packages.

The best way to start is to have a look at the vignette:

\code{vignette("seas")}
}
\author{
Christoph Sax \email{christoph.sax@gmail.com}
}
\seealso{
\code{\link{seas}} for more information on usage.
}
\keyword{package}

