% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{seasonal-package}
\alias{seasonal}
\alias{seasonal-package}
\title{seasonal: R interface to X-13ARIMA-SEATS}
\description{
seasonal is an easy-to-use and (almost) full-featured
  R-interface to X-13ARIMA-SEATS, the newest seasonal adjustment software
  developed by the United States Census Bureau. X-13ARIMA-SEATS combines and
  extends the capabilities of the older X-12ARIMA (developed by the Census
  Bureau) and TRAMO-SEATS (developed by the Bank of Spain).

  The best way to start is to have a look at the vignette, where installation
  and usage is described.

  \code{vignette("seas")}
}
\author{
Christoph Sax \email{christoph.sax@gmail.com}
}
\seealso{
\code{\link{seas}} for the core function and more information on
  package usage.
}
\keyword{package}

