\name{series}
\alias{series}
\title{Universal import of X-13 data series}
\usage{
series(x, series, reeval = TRUE)
}
\arguments{
  \item{x}{an object of class \code{"seas"}.}

  \item{series}{short name of an X-13 data table.}
}
\description{
Universal import of X-13 data series
}
\examples{
\dontrun{

}
}

