% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seasonal-package.R
\docType{data}
\name{SPECS}
\alias{SPECS}
\title{List of Available X-13ARIMA-SEATS Outputs}
\format{An object of class \code{"data.frame"}}
\source{
United States Census Bureau
}
\description{
The data is used by several functions as a look-up table. Users
should consider the table in \code{\link{series}} or in the official manual.
}
\references{
Official X-13ARIMA-SEATS manual:
  \url{http://www.census.gov/ts/x13as/docX13AS.pdf}
}
\keyword{datasets}

