% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{transformfunction}
\alias{transformfunction}
\title{Applied Transformation}
\usage{
transformfunction(x)
}
\arguments{
\item{x}{object of class \code{"seas"}}
}
\description{
Returns the transform function that has been applied.
}
\examples{
\donttest{

m <- seas(AirPassengers)
transformfunction(m)
}
}
\references{
Vignette with a more detailed description:
\url{http://www.seasonal.website/seasonal.html}

Comprehensive list of R examples from the X-13ARIMA-SEATS manual:
\url{http://www.seasonal.website/examples.html}

Official X-13ARIMA-SEATS manual:
\url{https://www2.census.gov/software/x-13arima-seats/x13as/windows/documentation/docx13as.pdf}
}
\seealso{
\code{\link[=seas]{seas()}} for the main function.

\code{\link[=series]{series()}}, for universal X-13 output extraction.

\code{\link[=plot.seas]{plot.seas()}}, for diagnostical plots.

\code{\link[=out]{out()}}, for accessing the full output of X-13ARIMA-SEATS.
}
