% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udg.R
\name{udg}
\alias{AIC.seas}
\alias{BIC.seas}
\alias{logLik.seas}
\alias{nobs.seas}
\alias{udg}
\title{X-13ARIMA-SEATS Stats}
\usage{
udg(x, stats = NULL, simplify = TRUE, fail = TRUE)

\method{AIC}{seas}(object, ...)

\method{BIC}{seas}(object, ...)

\method{nobs}{seas}(object, ...)

\method{logLik}{seas}(object, ...)
}
\arguments{
\item{x, object}{an object of class \code{"seas"}.}

\item{stats}{character vector. If specified, only a subset of the available 
stats are returned. This speeds up the call, as only a subset needs to be 
parsed. Should be used for programming.}

\item{simplify}{logical; should the result be simplified to a vector, if possible?}

\item{fail}{logical; if TRUE, an error is droped if an element of 
\code{stats} is missing in \code{names(udg(x))}.}

\item{...}{further arguments (unused)}
}
\value{
a named vector or list.
}
\description{
The udg file contains most statistics of an X-13 run. The 
\code{udg} function returns its content in a properly parsed form.
}
\examples{
\dontrun{
m <- seas(AirPassengers, x11 = "")

# a list with all entries from udg
udg(m)

# extracting a few selected stats from udg
udg(m, c("f3.m02", "f3.m05", "qsori"))  # returns a list
udg(m, c("f3.m02", "f3.m05"))           # returns a vector

# faster than:
udg(m)[c("f3.m01", "f3.m02", "qsori")]
}
}

