% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udg.R
\name{udg}
\alias{udg}
\title{X-13ARIMA-SEATS Stats}
\usage{
udg(x, stats = NULL, simplify = TRUE, fail = TRUE)
}
\arguments{
\item{x}{an object of class \code{"seas"}.}

\item{stats}{character vector. If specified, only a subset of the available 
stats are returned. This speeds up the call, as only a subset needs to be 
parsed. Should be used for programming.}

\item{simplify}{logical; should the result be simplified to a vector, if possible?}

\item{fail}{logical; if TRUE, an error is droped if an element of 
\code{stats} iis missing in the \code{names(udg(x))}.}
}
\value{
a named \code{"list"}.
}
\description{
Returns an object of class \code{"list"} that contains properly parsed
content of the \code{udg} stats file.
}
\examples{

m <- seas(AirPassengers, x11 = "")

# a list with all entries from udg
udg(m)
names(udg(m))  # show all available stats

# extracting a few selected stats from udg
udg(m, c("f3.m02", "f3.m05", "qsori"))  # returns a list
udg(m, c("f3.m02", "f3.m05"))           # returns a vector

# faster than:
udg(m)[c("f3.m01", "f3.m02", "qsori")]
}

