% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udg.R
\name{udg}
\alias{udg}
\title{X-13ARIMA-SEATS Stats}
\usage{
udg(x, stats = NULL, fail = TRUE)
}
\arguments{
\item{x}{an object of class \code{"seas"}.}

\item{stats}{character vector. If specified, only a subset of the available 
stats are returned. This speeds up the call, as only a subset needs to be 
parsed. Should be used for programming.}

\item{fail}{if TRUE, an error is droped if a stat is not in the \code{udg}
file.}
}
\value{
a named \code{"list"}.
}
\description{
Returns an object of class \code{"list"} that contains properly parsed
content of the \code{udg} stat file.
}
\examples{

m <- seas(AirPassengers, x11 = "")

# a list with all entries from udg
udg(m)
names(udg(m))  # show all available stats

# extracting a few selected stats from udg
udg(m, c("f3.m01", "f3.m02", "f3.m03"))

# this is faster than:
udg(m)[c("f3.m01", "f3.m02", "f3.m03")]

# a table with some m quality mesures
do.call(rbind, udg(m, c("f3.m01", "f3.m02", "f3.m03")))
}

